Restaurant Activity up

+++ src/com/gourmet6/Restaurant.java
@@ -24,7 +24,7 @@ public class Restaurant {
 	private String description;
 	private ArrayList<String> cuisines;
 	private ArrayList<String> photos;
-	private byte rating;
+	private float rating;
 	private short zip;
 	private short seats;
 	private short availableSeats; // en temps reel
@@ -49,7 +49,7 @@ public class Restaurant {
 	public Restaurant(String restaurant)
 	{
 		//TODO DB query based on the restaurant's name
-		// problme : pas possible sans contexte -> voir dans RestaurantDBHandler
+		// problï¿½me : pas possible sans contexte -> voir dans RestaurantDBHandler
 		
 	}
 	
@@ -157,7 +157,7 @@ public class Restaurant {
 	public void setPhotos(ArrayList<String> photos) {
 		this.photos = photos;
 	}
-	public byte getRating() {
+	public float getRating() {
 		return rating;
 	}
 	public void setRating(byte rating) {
@@ -221,9 +221,9 @@ public class Restaurant {
 	 * 
 	 * @param res
 	 * @return cette methode renvoit true si la reservation est correcte, c-a-d:
-	 * -Qu'elle specifie des plats qui sont encore disponibles, si une commande est joind à la reservation
+	 * -Qu'elle specifie des plats qui sont encore disponibles, si une commande est joind Ã  la reservation
 	 * -Qu'elle reserve lorsque le restaurant est ouvert.
-	 * -Qu'elle spécifie une nombre de personne <= au nombre de place encore disponible dans le restaurant.
+	 * -Qu'elle spÃ©cifie une nombre de personne <= au nombre de place encore disponible dans le restaurant.
 	 * Renvoit false sinon.
 	 */
 	public boolean checkReservation(Reservation res){
